+++ FitnessTracker.Models/Exercise/ExerciseDetail.cs
@@ -17,7 +17,7 @@ namespace FitnessTracker.Models
         [Display(Name = "Name")]
         public string NameOfExercise { get; set; }
         [Required]
-        public float Duration { get; set; }
+        public int Duration { get; set; }
         [Required]
         [Display(Name = "Type")]
         public ExerciseType TypeOfExercise { get; set; }
